Se agrega validacion en MovimientoUseCase
